requirements.txt
# Enhanced AI Trading System Requirements
# Core dependencies
pandas>=2.0.0
numpy>=1.24.0
yfinance>=0.2.0
requests>=2.31.0
python-dotenv>=1.0.0

# Database
sqlalchemy>=2.0.0
sqlite3

# Data validation
pydantic>=2.0.0
marshmallow>=3.19.0

# Visualization
matplotlib>=3.7.0
plotly>=5.15.0
seaborn>=0.12.0

# Logging and monitoring
loguru>=0.7.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0

# API and HTTP
httpx>=0.24.0
aiohttp>=3.8.0

# Date/time handling
python-dateutil>=2.8.0
pytz>=2023.3

# Financial calculations
TA-Lib>=0.4.26
quantlib>=1.31

# Configuration
click>=8.1.0
toml>=0.10.2

# Development
black>=23.0.0
flake8>=6.0.0
mypy>=1.4.0
